refactor: 💡 Rename collection item
