+++ apps/webstore/src/app/pages/auth/collection/collection.component.spec.ts
@@ -1,6 +1,6 @@
 import { HttpClientModule } from '@angular/common/http';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { FamilyTreeCollectionItemComponent } from '../../../shared/components/items/family-tree-collection-item/family-tree-collection-item.component';
+import { FamilyTreeCollectionItemComponent } from '../../../shared/components/items/collection-item/collection-item.component';
 import { CollectionComponent } from './collection.component';
 
 describe('CollectionComponent', () => {
