+++ apps/webstore/src/app/shared/components/items/collection-item/collection-item.component.ts
@@ -1,32 +1,28 @@
 import { HttpErrorResponse } from '@angular/common/http';
-import { Component, EventEmitter, Input, Output, ViewChild } from '@angular/core';
-import { Router } from '@angular/router';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { DesignTypeEnum, IDesign, IFamilyTree } from '@interfaces';
 import { LocalStorageService } from '@local-storage';
 import { LocalStorageVars } from '@models';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { AddToBasketModalComponent } from '../../../../shared/components/modals/add-to-basket-modal/add-to-basket-modal.component';
+import { AddToBasketModalComponent } from '../../modals/add-to-basket-modal/add-to-basket-modal.component';
 import { DesignService } from '../../../services/design/design.service';
-import { FamilyTreeDesignComponent } from '../../products/family-tree/family-tree-design/family-tree-design.component';
 import { ToastService } from '../../toast/toast-service';
 
 @Component({
-  selector: 'webstore-family-tree-collection-item',
-  templateUrl: './family-tree-collection-item.component.html',
-  styleUrls: ['./family-tree-collection-item.component.css'],
+  selector: 'webstore-collection-item',
+  templateUrl: './collection-item.component.html',
+  styleUrls: ['./collection-item.component.css'],
 })
-export class FamilyTreeCollectionItemComponent {
-  @ViewChild('productDesign', { static: true })
-  miniature: FamilyTreeDesignComponent;
+export class CollectionItemComponent {
   @Input() design: IDesign;
   isLoading = false;
+
   @Output() deleteEvent = new EventEmitter();
   public designTypeEnum = DesignTypeEnum;
 
   constructor(
     private toastService: ToastService,
     private designService: DesignService,
-    private router: Router,
     private modalService: NgbModal,
     private localStorageService: LocalStorageService
   ) {}
